+++ lib/twofold.dart
@@ -1,4 +1,2 @@
-export 'src/async/twofold_future_extensions.dart';
-export 'src/async/twofold_future_helpers.dart';
 export 'src/core/twofold.dart';
 export 'src/testing/twofold_expect.dart';